pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-financial-analyzer-v1"
version = "1.0.0"
description = "AI-powered financial statement analysis CLI tool"
authors = [
    { name="Pranav Khatade" }
]
dependencies = [
    "google-genai",
    "matplotlib",
    "reportlab"
]
requires-python = ">=3.9"

[project.scripts]
financial-analyzer = "app.cli:main"

[tool.setuptools.packages.find]
include = ["app*"]